Spread large file into smaller chunks with less responsibilities
