+++ internal/submodelrepository/persistence/submodel_file_attachment_database.go
@@ -1,0 +1,127 @@
+/*******************************************************************************
+* Copyright (C) 2026 the Eclipse BaSyx Authors and Fraunhofer IESE
+*
+* Permission is hereby granted, free of charge, to any person obtaining
+* a copy of this software and associated documentation files (the
+* "Software"), to deal in the Software without restriction, including
+* without limitation the rights to use, copy, modify, merge, publish,
+* distribute, sublicense, and/or sell copies of the Software, and to
+* permit persons to whom the Software is furnished to do so, subject to
+* the following conditions:
+*
+* The above copyright notice and this permission notice shall be
+* included in all copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+* EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+* MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+* NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+* LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+* OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+* WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+*
+* SPDX-License-Identifier: MIT
+******************************************************************************/
+// Author: Jannik Fried (Fraunhofer IESE)
+
+package persistence
+
+import (
+	"context"
+	"database/sql"
+	"errors"
+	"os"
+
+	"github.com/eclipse-basyx/basyx-go-components/internal/common"
+	submodelqueries "github.com/eclipse-basyx/basyx-go-components/internal/submodelrepository/persistence/queries"
+	submodelelements "github.com/eclipse-basyx/basyx-go-components/internal/submodelrepository/persistence/submodelElements"
+)
+
+// FileAttachmentExists reports whether a File submodel element currently has attachment data stored in file_data.file_oid.
+func (s *SubmodelDatabase) FileAttachmentExists(submodelID string, idShortPath string) (bool, error) {
+	query, args, err := submodelqueries.BuildFileAttachmentExistsSQL(submodelID, idShortPath)
+	if err != nil {
+		return false, common.NewInternalServerError("SMREPO-FILEATTEXISTS-BUILDSQL " + err.Error())
+	}
+
+	var fileElementID sql.NullInt64
+	var fileOID sql.NullInt64
+	if scanErr := s.db.QueryRow(query, args...).Scan(&fileElementID, &fileOID); scanErr != nil {
+		if errors.Is(scanErr, sql.ErrNoRows) {
+			return false, common.NewErrNotFound("SMREPO-FILEATTEXISTS-NOTFOUND Submodel element not found")
+		}
+		return false, common.NewInternalServerError("SMREPO-FILEATTEXISTS-QUERY " + scanErr.Error())
+	}
+
+	if !fileElementID.Valid {
+		return false, common.NewErrMethodNotAllowed("SMREPO-FILEATTEXISTS-NOTFILE Submodel element is not of type File")
+	}
+
+	return fileOID.Valid, nil
+}
+
+// UploadFileAttachment uploads attachment content for a File submodel element.
+func (s *SubmodelDatabase) UploadFileAttachment(submodelID string, idShortPath string, file *os.File, fileName string) error {
+	fileHandler, err := submodelelements.NewPostgreSQLFileHandler(s.db)
+	if err != nil {
+		return err
+	}
+
+	return fileHandler.UploadFileAttachment(submodelID, idShortPath, file, fileName)
+}
+
+// UploadFileAttachmentWithHistory uploads attachment content and appends the current Submodel snapshot atomically.
+func (s *SubmodelDatabase) UploadFileAttachmentWithHistory(ctx context.Context, submodelID string, idShortPath string, file *os.File, fileName string) error {
+	fileHandler, err := submodelelements.NewPostgreSQLFileHandler(s.db)
+	if err != nil {
+		return err
+	}
+
+	return common.ExecuteInTransaction(s.db, "SMREPO-UPLOADFILEHIST-STARTTX", "SMREPO-UPLOADFILEHIST-COMMIT", func(tx *sql.Tx) error {
+		if err := fileHandler.UploadFileAttachmentTx(tx, submodelID, idShortPath, file, fileName); err != nil {
+			return err
+		}
+		return s.appendChangedSubmodelElementHistoryTx(ctx, tx, submodelID, submodelElementRootMutation{
+			previousPath: idShortPath,
+			currentPath:  idShortPath,
+		})
+	})
+}
+
+// DownloadFileAttachment downloads attachment content for a File submodel element.
+func (s *SubmodelDatabase) DownloadFileAttachment(submodelID string, idShortPath string) ([]byte, string, string, error) {
+	fileHandler, err := submodelelements.NewPostgreSQLFileHandler(s.db)
+	if err != nil {
+		return nil, "", "", err
+	}
+
+	return fileHandler.DownloadFileAttachment(submodelID, idShortPath)
+}
+
+// DeleteFileAttachment deletes attachment content of a File submodel element.
+func (s *SubmodelDatabase) DeleteFileAttachment(submodelID string, idShortPath string) error {
+	fileHandler, err := submodelelements.NewPostgreSQLFileHandler(s.db)
+	if err != nil {
+		return err
+	}
+
+	return fileHandler.DeleteFileAttachment(submodelID, idShortPath)
+}
+
+// DeleteFileAttachmentWithHistory deletes attachment content and appends the current Submodel snapshot atomically.
+func (s *SubmodelDatabase) DeleteFileAttachmentWithHistory(ctx context.Context, submodelID string, idShortPath string) error {
+	fileHandler, err := submodelelements.NewPostgreSQLFileHandler(s.db)
+	if err != nil {
+		return err
+	}
+
+	return common.ExecuteInTransaction(s.db, "SMREPO-DELETEFILEHIST-STARTTX", "SMREPO-DELETEFILEHIST-COMMIT", func(tx *sql.Tx) error {
+		if err := fileHandler.DeleteFileAttachmentTx(tx, submodelID, idShortPath); err != nil {
+			return err
+		}
+		return s.appendChangedSubmodelElementHistoryTx(ctx, tx, submodelID, submodelElementRootMutation{
+			previousPath: idShortPath,
+			currentPath:  idShortPath,
+		})
+	})
+}
